requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.37.2
torch==2.2.0
python-pptx==0.6.23
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.0.1
Pillow==10.2.0
PyMuPDF==1.23.26
requests==2.31.0
